[update] 付款金額、項目通用性處理

+++ src/EtheriT.Coker.Application.Shared/Dto/Order/OrderRepaySetDto.cs
@@ -4,7 +4,7 @@ namespace EtheriT.Coker.Application.Shared.Dto.Order
     public class OrderRepaySetDto
     {
         public long ohid {  get; set; }
-        public int Subtotal { get; set; }
+        public decimal Subtotal { get; set; }
         public List<OrderRepaySetDetailsDto>? Details { get; set; }
     }
 }
